fix: make focus restoration work with slotted grid
